+++ src/parser/inline/link.test.ts
@@ -11,16 +11,10 @@ describe('Unit: parser/inline/link', () => {
       assert.deepStrictEqual(inspect(parser('[](vbscript:alert)')), undefined);
       assert.deepStrictEqual(inspect(parser('[](data:text/html;base64,PHNjcmlwdD5hbGVydCgnWFNTJyk8L3NjcmlwdD4K)')), undefined);
       assert.deepStrictEqual(inspect(parser('[](any:alert)')), undefined);
-      assert.deepStrictEqual(inspect(parser('[](/")')), [['<a href="/&quot;" rel="noopener">/"</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](/\\)')), undefined);
+      assert.deepStrictEqual(inspect(parser('[](")')), [['<a href="&quot;" rel="noopener">"</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[](\\)')), [['<a href="\\" rel="noopener">\\</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[](")')), [['<a href="&quot;" rel="noopener">"</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[]("#)')), [['<a href="&quot;#" rel="noopener">"#</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[]("\\)')), undefined);
-      assert.deepStrictEqual(inspect(parser('[](\\)')), undefined);
-      assert.deepStrictEqual(inspect(parser('[](\\")')), [['<a href="&quot;" rel="noopener">"</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](\\\\)')), [['<a href="\\" rel="noopener">\\</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](\\\\#)')), [['<a href="\\#" rel="noopener">\\#</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](\\\\")')), [['<a href="\\&quot;" rel="noopener">\\"</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[<img>](#)')), [['<a href="#" rel="noopener">&lt;img&gt;</a>'], '']);
     });
 
@@ -45,7 +39,6 @@ describe('Unit: parser/inline/link', () => {
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
-      assert.deepStrictEqual(inspect(parser('[]( )')), undefined);
       assert.deepStrictEqual(inspect(parser('[](# )')), undefined);
       assert.deepStrictEqual(inspect(parser('[]( #)')), undefined);
       assert.deepStrictEqual(inspect(parser('[]( # )')), undefined);
@@ -67,10 +60,12 @@ describe('Unit: parser/inline/link', () => {
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('[]()')), [[`<a href="" rel="noopener">${location.href}</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('[]( )')), [[`<a href="" rel="noopener">${location.href}</a>`], '']);
       assert.deepStrictEqual(inspect(parser('[](b)')), [['<a href="b" rel="noopener">b</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](b\\ )')), [['<a href="b%20" rel="noopener">b%20</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](\\ )')), [[`<a href="%20" rel="noopener">%20</a>`], '']);
-      assert.deepStrictEqual(inspect(parser('[](?b=c+d&\\ #)')), [['<a href="?b=c+d&amp;%20#" rel="noopener">?b=c+d&amp;%20#</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[](\\)')), [[`<a href="\\" rel="noopener">\\</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('[](\\ )')), [[`<a href="\\" rel="noopener">\\</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('[](\\b)')), [[`<a href="\\b" rel="noopener">\\b</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('[](?b=c+d&\\#)')), [['<a href="?b=c+d&amp;\\#" rel="noopener">?b=c+d&amp;\\#</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[](?&amp;)')), [['<a href="?&amp;amp;" rel="noopener">?&amp;amp;</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[a]()')), [['<a href="" rel="noopener">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[a](b)')), [['<a href="b" rel="noopener">a</a>'], '']);
@@ -98,9 +93,10 @@ describe('Unit: parser/inline/link', () => {
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('[](/())')), [['<a href="/()" rel="noopener">/()</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](/(()))')), [['<a href="/(())" rel="noopener">/(())</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[](/(a(b)()(c)d))')), [['<a href="/(a(b)()(c)d)" rel="noopener">/(a(b)()(c)d)</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[a](())')), [['<a href="" rel="noopener">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[a](()))')), [['<a href="" rel="noopener">a</a>'], ')']);
+      assert.deepStrictEqual(inspect(parser('[a](() ))')), [['<a href=")" rel="noopener">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[[](/)](/)')), undefined);
       assert.deepStrictEqual(inspect(parser('[((/))](/)')), undefined);
       assert.deepStrictEqual(inspect(parser('[<wbr>](/)')), undefined);
@@ -108,8 +104,6 @@ describe('Unit: parser/inline/link', () => {
       assert.deepStrictEqual(inspect(parser('[!http://host](/)')), [['<a href="/" rel="noopener" target="_blank"><img class="media" data-src="http://host" alt=""></a>'], '']);
       assert.deepStrictEqual(inspect(parser('[\\[](/)')), [['<a href="/" rel="noopener">[</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[\\]](/)')), [['<a href="/" rel="noopener">]</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](/\\()')), [['<a href="/(" rel="noopener">/(</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](/\\))')), [['<a href="/)" rel="noopener">/)</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[*a*](b)')), [['<a href="b" rel="noopener"><em>a</em></a>'], '']);
       assert.deepStrictEqual(inspect(parser('[<wbr>"]("?"#")')), [['<a href="&quot;?&quot;#&quot;" rel="noopener"><wbr>"</a>'], '']);
     });
@@ -120,11 +114,10 @@ describe('Unit: parser/inline/link', () => {
     });
 
     it('nofollow', () => {
-      assert.deepStrictEqual(inspect(parser('[](/ nofollow )')), undefined);
-      assert.deepStrictEqual(inspect(parser('[](/  nofollow)')), undefined);
+      assert.deepStrictEqual(inspect(parser('[]( nofollow )')), undefined);
+      assert.deepStrictEqual(inspect(parser('[](  nofollow)')), undefined);
       assert.deepStrictEqual(inspect(parser('[]( nofollow)')), [[`<a href="" rel="noopener nofollow noreferrer">${location.href.slice(1)}</a>`], '']);
       assert.deepStrictEqual(inspect(parser('[](/ nofollow)')), [['<a href="/" rel="noopener nofollow noreferrer">/</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[](\\  nofollow)')), [['<a href="%20" rel="noopener nofollow noreferrer">%20</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[](http://host nofollow)')), [['<a href="http://host" rel="noopener nofollow noreferrer" target="_blank">ttp://host</a>'], '']);
     });