+++ src/parser/inline/media.test.ts
@@ -11,17 +11,8 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![](vbscript:alert)')), undefined);
       assert.deepStrictEqual(inspect(parser('![](data-type="image" data:text/html;base64,PHNjcmlwdD5hbGVydCgnWFNTJyk8L3NjcmlwdD4K)')), undefined);
       assert.deepStrictEqual(inspect(parser('![](any:alert)')), undefined);
-      assert.deepStrictEqual(inspect(parser('![](#")')), [['<img class="media" data-src="#&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](#\\)')), undefined);
       assert.deepStrictEqual(inspect(parser('![](")')), [['<img class="media" data-src="&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![]("#)')), [['<img class="media" data-src="&quot;#" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![]("\\)')), undefined);
-      assert.deepStrictEqual(inspect(parser('![](\\)')), undefined);
-      assert.deepStrictEqual(inspect(parser('![](\\#)')), [['<img class="media" data-src="#" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\")')), [['<img class="media" data-src="&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\)')), [['<img class="media" data-src="\\" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\#)')), [['<img class="media" data-src="\\#" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\")')), [['<img class="media" data-src="\\&quot;" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\)')), [['<img class="media" data-src="\\" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('!["](#)')), [['<img class="media" data-src="#" alt="&quot;">'], '']);
       assert.deepStrictEqual(inspect(parser('![\\"](#)')), [['<img class="media" data-src="#" alt="&quot;">'], '']);
     });
@@ -33,7 +24,6 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![]( )')), undefined);
       assert.deepStrictEqual(inspect(parser('![ ]( )')), undefined);
       assert.deepStrictEqual(inspect(parser('![]( #)')), undefined);
-      assert.deepStrictEqual(inspect(parser('![](# )')), undefined);
       assert.deepStrictEqual(inspect(parser('![]( # )')), undefined);
       assert.deepStrictEqual(inspect(parser('![](# #)')), undefined);
       assert.deepStrictEqual(inspect(parser('![]](#)')), undefined);
@@ -53,8 +43,10 @@ describe('Unit: parser/inline/media', () => {
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('![](b)')), [['<img class="media" data-src="b" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](b\\ )')), [['<img class="media" data-src="b%20" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](b\\ c)')), [['<img class="media" data-src="b%20c" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](b )')), [['<img class="media" data-src="b" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\)')), [['<img class="media" data-src="\\" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\ )')), [['<img class="media" data-src="\\" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\b)')), [['<img class="media" data-src="\\b" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![a](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
       assert.deepStrictEqual(inspect(parser('![a b](c)')), [['<img class="media" data-src="c" alt="a b">'], '']);
       assert.deepStrictEqual(inspect(parser('![a ](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
@@ -63,9 +55,10 @@ describe('Unit: parser/inline/media', () => {
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('![](#())')), [['<img class="media" data-src="#()" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](#(()))')), [['<img class="media" data-src="#(())" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![](#(a(b)()(c)d))')), [['<img class="media" data-src="#(a(b)()(c)d)" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![]((b))')), [['<img class="media" data-src="b" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![]((b)))')), [['<img class="media" data-src="b" alt="">'], ')']);
+      assert.deepStrictEqual(inspect(parser('![]((b) ))')), [['<img class="media" data-src="b)" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![\\[](#)')), [['<img class="media" data-src="#" alt="[">'], '']);
       assert.deepStrictEqual(inspect(parser('![<wbr>"]("?"#")')), [['<img class="media" data-src="&quot;?&quot;#&quot;" alt="<wbr>&quot;">'], '']);
     });